% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_diff_result}
\alias{example_diff_result}
\title{Output of running paired_diff on example_se.}
\format{
A `DataFrame` with 954 rows and 10 columns.
}
\usage{
data("example_diff_result")
}
\value{
A `DataFrame`.
}
\description{
This example result is used primarily to do package tests
and for function man pages
}
\keyword{datasets}
