% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gene_sets}
\alias{example_gene_sets}
\title{MSigDB gene sets from humans, category C5 with ensemble gene IDs}
\format{
A list of 77 human gene sets
}
\usage{
data("example_gene_sets")
}
\description{
This example gene set list is used primarily to do package tests
and for function man pages.
}
\keyword{datasets}
