% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ora_results}
\alias{example_ora_results}
\title{Output of running paired_ora on example_diff_result
and gene sets extracted from MSigDB}
\format{
A data.frame with 559 rows and 18 columns.
}
\usage{
data("example_ora_results")
}
\description{
This example result is used primarily to do package tests
and for function man pages.
}
\keyword{datasets}
