% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_se}
\alias{example_se}
\title{A small subset of the GEO:GSE61220 data set.}
\format{
A `SummarizedExperiment`
\describe{
    \item{assay}{Count matrix with 5611 transcripts and 6 samples}
    \item{colData}{The metadata associated with the count matrix}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE61220}
}
\usage{
data("example_se")
}
\value{
A `SummarizedExperiment`
}
\description{
The subset is used in the vignettes and function man pages.
The subset was created by extracting genes belonging to
Telomere-related gene sets and randomly selecting
900 other genes from the original dataset.
}
\keyword{datasets}
