% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{formularise_vector}
\alias{formularise_vector}
\title{Convert character vector into a stats formula}
\usage{
formularise_vector(vector)
}
\arguments{
\item{vector}{Character vector of variables to add to model}
}
\description{
Convert character vector into a stats formula
}
