% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getDDS}
\alias{getDDS}
\title{Run SVA and export dds}
\usage{
getDDS(
  row,
  archs4db = NULL,
  tx_count = NULL,
  group_col = "group_nr",
  tpm = TRUE,
  prefilter = 10,
  parallel = TRUE
)
}
\description{
Run SVA and export dds
}
