% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_de.R
\name{paired_de}
\alias{paired_de}
\title{Run paired DESeq2 and DEXSeq analyses}
\usage{
paired_de(
  tx_count,
  metadata,
  group_col,
  sample_col,
  baseline,
  case,
  covariates = NULL,
  experiment_title = "Experiment Title",
  store_results = TRUE,
  run_sva = TRUE,
  prefilter = 10,
  test = "LRT",
  fit_type = "local",
  quiet = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam(),
  ...
)
}
\arguments{
\item{tx_count}{The transcripts count matrix of an RNA-seq analysis}

\item{metadata}{A metadata file or data frame object}

\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{sample_col}{The column in the metadata that specifies the sample IDs (should correspond to column names in tx_count)}

\item{baseline}{Group value of baseline samples}

\item{case}{Group value of case samples}

\item{covariates}{Name of column(s) in the metadata that indicate(s) covariates. E.g., c("gender", "tissue_type")}

\item{experiment_title}{Title of your experiment. Your results will be stored in paste0("results/", experiment_title, "_pairedGSEA.RDS").}

\item{store_results}{(Default: TRUE) A logical indicating if results should be stored in the folder "results/".}

\item{run_sva}{(Default: TRUE) A logical stating whether SVA should be run.}

\item{prefilter}{(Default: 10) The prefilter threshold, where rowSums lower than the prefilter threshold will be removed from the count matrix. Set to 0 or FALSE to prevent prefiltering}

\item{test}{either "Wald" or "LRT", which will then use either 
Wald significance tests (defined by \code{\link[DESeq2]{nbinomWaldTest}}),
or the likelihood ratio test on the difference in deviance between a
full and reduced model formula (defined by \code{\link[DESeq2]{nbinomLRT}})}

\item{fit_type}{(Default: "local") Either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity.}

\item{quiet}{(Default: FALSE) Whether to print messages}

\item{parallel}{(Default: FALSE) If FALSE, no parallelization. If TRUE, parallel execution using BiocParallel, see next argument BPPARAM.}

\item{BPPARAM}{(Default: \code{BiocParallel::bpparam()}) An optional parameter object passed internally to bplapply when parallel = TRUE. If not specified, the parameters last registered with register will be used.}

\item{...}{Additional parameters passed to \code{\link[DESeq2:DESeq]{DESeq()}}}
}
\description{
Run paired DESeq2 and DEXSeq analyses
}
\seealso{
Other paired: 
\code{\link{paired_ora}()}
}
\concept{paired}
