% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_gsea.R
\name{paired_gsea}
\alias{paired_gsea}
\title{Run paired DESeq2 and DEXSeq analyses}
\usage{
paired_gsea(
  tx_count,
  metadata,
  group_col,
  sample_col,
  comparison,
  experiment_title = "MyLittleExperiment",
  run_sva = TRUE,
  prefilter = 10,
  fit_type = "local",
  dds_out = FALSE,
  dxd_out = FALSE,
  quiet = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{tx_count}{The transcripts count matrix of an RNA-seq analysis}

\item{metadata}{A metadata file or data frame object}

\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{sample_col}{The column in the metadata that specifies the sample IDs (should correspond to column names in tx_count)}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}

\item{experiment_title}{Title of your experiment. Your results will be stored in paste0("results/", experiment_title, "_pairedGSEA.RDS").}

\item{run_sva}{(Default: TRUE) A logical stating whether SVA should be run.}

\item{prefilter}{(Default: 10) The prefilter threshold, where rowSums lower than the prefilter threshold will be removed from the count matrix. Set to 0 or FALSE to prevent prefiltering}

\item{fit_type}{(Default: "local") Either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity.}

\item{dds_out}{(Default: FALSE) Put a file string to store DESeqDataSet after DESeq analysis, before results are extracted}

\item{dxd_out}{(Default: FALSE) Put a file string to store DEXSeqDataSet after DEXSeq analysis, before results are extracted}

\item{quiet}{(Default: FALSE) Whether to print messages}

\item{parallel}{(Default: FALSE) If FALSE, no parallelization. If TRUE, parallel execution using BiocParallel, see next argument BPPARAM.}

\item{BPPARAM}{(Default: \code{BiocParallel::bpparam()}) An optional parameter object passed internally to bplapply when parallel = TRUE. If not specified, the parameters last registered with register will be used.}
}
\description{
Run paired DESeq2 and DEXSeq analyses
}
