% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEXSeq.R
\name{perGenePValue}
\alias{perGenePValue}
\title{Per gene p value aggregation}
\usage{
perGenePValue(
  df,
  gene = "gene",
  p = "pvalue",
  weights = "baseMean",
  lfc = NULL
)
}
\description{
Per gene p value aggregation
}
