% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Prepare metadata}
\usage{
prepare_metadata(metadata, group_col, comparison)
}
\arguments{
\item{metadata}{A metadata file or data frame object}

\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}
}
\description{
Prepare metadata
}
