% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{prepare_tx_count}
\alias{prepare_tx_count}
\title{Prepare DESeqDataSet from ARCHS4 database}
\usage{
prepare_tx_count(
  metadata,
  group_col,
  comparison,
  archs4db = NULL,
  gtf = NULL,
  samples = "id"
)
}
\arguments{
\item{metadata}{A metadata file or data frame object}

\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}

\item{archs4db}{The path to an ARCHS4 database file}

\item{gtf}{Optional: Adds GTF gene and transcript labels to the ARCHS4 data extract}

\item{samples}{The column in the metadata that specifies the samples}
}
\description{
Prepare DESeqDataSet from ARCHS4 database
}
