% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{prepMeta}
\alias{prepMeta}
\title{Prepare metadata}
\usage{
prepMeta(md, groupCol, comparison)
}
\arguments{
\item{md}{A metadata file or data frame object}

\item{groupCol}{The metadata column specifying the what group each sample is associated with}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}
}
\description{
Prepare metadata
}
