% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{reduce_formula}
\alias{reduce_formula}
\title{Reduce design formula by removing first variable}
\usage{
reduce_formula(formula, formularise = TRUE)
}
\arguments{
\item{formula}{A design formula to reduce. The first variable will be removed.}

\item{formularise}{(Default: TRUE) A logical determining if the design should be formularised}
}
\description{
This internal function assumes the design formula has the format as imposed by \code{\link[pairedGSEA:formularise_vector]{formularise_vector()}}
  and that the defining variable is the first occurring variable in the design.
}
