% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dexseq.R
\name{run_dexseq}
\alias{run_dexseq}
\title{Prepare DEXSeqDataSet from DESeqDataSet}
\usage{
run_dexseq(
  dds,
  group_col,
  comparison,
  dxd_out = FALSE,
  quiet = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}

\item{dxd_out}{(Default: FALSE) Put a file string to store DEXSeqDataSet after DEXSeq analysis, before results are extracted}

\item{quiet}{(Default: FALSE) Whether to print messages}

\item{parallel}{(Default: FALSE) If FALSE, no parallelization. If TRUE, parallel execution using BiocParallel, see next argument BPPARAM.}

\item{BPPARAM}{(Default: \code{BiocParallel::bpparam()}) An optional parameter object passed internally to bplapply when parallel = TRUE. If not specified, the parameters last registered with register will be used.}
}
\description{
Prepare DEXSeqDataSet from DESeqDataSet
}
