% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_diff.R
\name{run_limma}
\alias{run_limma}
\title{Run limma and voom analyses}
\usage{
run_limma(
    dds,
    group_col,
    baseline,
    case,
    experiment_title = "Experiment Title",
    store_results = FALSE,
    quiet = FALSE
    )
}
\arguments{
\item{dds}{A DESeqDataSet. See \code{\link[DESeq2]{DESeqDataSet}} 
for more information about the object type.}

\item{group_col}{The metadata column specifying the what group each
sample is associated with}

\item{baseline}{Group value of baseline samples}

\item{case}{Group value of case samples}

\item{experiment_title}{Title of your experiment. Your results will be
stored in \code{paste0("results/", experiment_title, "_pairedGSEA.RDS")}.}

\item{store_results}{(Default: \code{FALSE})
A logical indicating if results should
be stored in the folder \code{"results/"}.}

\item{quiet}{(Default: \code{FALSE}) Whether to print messages}
}
\value{
A DFrame
}
\description{
This internal function runs a differential gene expression analysis using
limma (See \code{\link[limma]{lmFit}} for more detailed information), and 
differential gene splicing using voom (See \code{\link[limma]{voom}}).
Here, the surrogate variables found by \code{\link{run_sva}},
if any, will be added to the DESeqDataSet before running the analysis.
}
\keyword{internal}
