% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Run DESeq2}
\usage{
runDESeq2(
  dds,
  groupCol,
  comparison,
  fitType = "local",
  tpm = FALSE,
  samples = NULL,
  gtf = NULL,
  dds_out = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{groupCol}{The metadata column specifying the what group each sample is associated with}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}

\item{samples}{The column in the metadata that specifies the samples}

\item{gtf}{Optional: Adds GTF gene and transcript labels to the ARCHS4 data extract}
}
\description{
Run DESeq2
}
