% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{runExperiment}
\alias{runExperiment}
\title{Run DESeq2 and DEXSeq analyses}
\usage{
runExperiment(
  row,
  archs4db = NULL,
  tx_count = NULL,
  group_col = "group_nr",
  tpm = TRUE,
  prefilter = 10,
  parallel = TRUE
)
}
\description{
Run DESeq2 and DEXSeq analyses
}
