% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{runSVA}
\alias{runSVA}
\title{Run SVA}
\usage{
runSVA(txCount, metadata, groupCol)
}
\arguments{
\item{txCount}{The count matrix of an RNA-seq analysis}

\item{groupCol}{The metadata column specifying the what group each sample is associated with}
}
\description{
Run SVA
}
