% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ora.R
\name{spearman}
\alias{spearman}
\title{Compute Spearman correlation between enrichment scores}
\usage{
spearman(ora)
}
\arguments{
\item{ora}{Output of \code{\link[pairedGSEA:paired_ora]{paired_ora}}}
}
\value{
A numeric value indicating the
Spearman correlation between the enrichment scores.
}
\description{
Compute Spearman correlation between enrichment scores
}
\keyword{internal}
