% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_diff.R
\name{aggregate_pvalue}
\alias{aggregate_pvalue}
\title{p-value aggregation to gene level}
\usage{
aggregate_pvalue(
    df,
    gene = "gene",
    p = "pvalue",
    weights = "baseMean",
    lfc = "lfc",
    type = "expression"
    )
}
\arguments{
\item{df}{A data.frame-type object with pvalues to aggregate}

\item{gene}{(Default: "gene") The column with gene names}

\item{p}{(Default: "pvalue") The column with p-values}

\item{weights}{(Default: "baseMean") The column to use
for weighting the aggregation}

\item{lfc}{(Default: "lfc") The column with log2-foldchanges}

\item{type}{(Default: "expression") 
Either "expression" for aggregation of
DESeq2 results or "splicing" for aggregation of DEXSeq results}
}
\value{
A data.frame
}
\description{
This internal function aggregates p-values to gene level using
Lancaster aggregation.
}
\keyword{internal}
