% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_diff.R
\name{convert_matrix_to_dds}
\alias{convert_matrix_to_dds}
\title{Convert matrix to \code{\link[DESeq2]{DESeqDataSet}}}
\usage{
convert_matrix_to_dds(tx_count, metadata, design)
}
\arguments{
\item{tx_count}{Count matrix to convert to dds}

\item{metadata}{(Default: \code{NULL}) A metadata file or
\code{data.frame} object}
}
\value{
A \code{\link[DESeq2]{DESeqDataSet}}
}
\description{
This internal function converts a matrix and metadata object into a
\code{\link[DESeq2]{DESeqDataSet}}, using the \code{group_col}
as the basic design.
}
\keyword{internal}
