% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{perGenePValue}
\alias{perGenePValue}
\title{Per gene p value aggregation}
\usage{
perGenePValue(
  df,
  gene = "gene",
  p = "pvalue",
  weights = "baseMean",
  lfc = "log2FC",
  type = "deseq2"
)
}
\description{
Per gene p value aggregation
}
