% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ora.R
\name{plot_ora}
\alias{plot_ora}
\title{Scatter plot of Over-Representation Analysis results}
\usage{
plot_ora(ora, plotly = FALSE, pattern = NULL)
}
\arguments{
\item{ora}{Output of \code{\link[pairedGSEA:paired_ora]{paired_ora()}}}

\item{plotly}{(Default: FALSE) Logical on whether to return plot as an interactive plotly plot or a simple ggplot.}

\item{pattern}{Highlight pathways containing a specific regex pattern}
}
\description{
Scatter plot of Over-Representation Analysis results
}
