% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{prepare_msigdb}
\alias{prepare_msigdb}
\title{Load MSigDB and convert to names list of gene sets}
\usage{
prepare_msigdb(category = "C5")
}
\arguments{
\item{category}{The MSigDB category to extract}
}
\description{
Load MSigDB and convert to names list of gene sets
}
