% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{prepDE}
\alias{prepDE}
\title{Pepare DE analysis}
\usage{
prepDE(
  md,
  archs4db,
  groupCol,
  comparison,
  tpm = FALSE,
  samples = "id",
  prefilter = 10,
  parallel = FALSE,
  cores = 4
)
}
\description{
Pepare DE analysis
}
