% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_diff.R
\name{run_deseq}
\alias{run_deseq}
\title{Run DESeq2 analysis}
\usage{
run_deseq(
    dds,
    group_col,
    baseline,
    case,
    test = "LRT",
    fit_type = "local",
    experiment_title = "Experiment Title",
    store_results = FALSE,
    quiet = FALSE,
    parallel = FALSE,
    BPPARAM = BiocParallel::bpparam(),
    ...
    )
}
\arguments{
\item{dds}{A DESeqDataSet. See \code{\link[DESeq2]{DESeqDataSet}} 
for more information about the object type.}

\item{group_col}{The metadata column specifying the what group each
sample is associated with}

\item{baseline}{Group value of baseline samples}

\item{case}{Group value of case samples}

\item{test}{either "Wald" or "LRT", which will then use either 
Wald significance tests (defined by \code{\link[DESeq2]{nbinomWaldTest}}),
or the likelihood ratio test on the difference in deviance between a
full and reduced model formula (defined by \code{\link[DESeq2]{nbinomLRT}})}

\item{fit_type}{(Default: \code{"local"}) Either 
\code{"parametric", "local", "mean", or "glmGamPoi"}
for the type of fitting of dispersions to the mean intensity.}

\item{experiment_title}{Title of your experiment. Your results will be
stored in \code{paste0("results/", experiment_title, "_pairedGSEA.RDS")}.}

\item{store_results}{(Default: \code{FALSE})
A logical indicating if results should
be stored in the folder \code{"results/"}.}

\item{quiet}{(Default: \code{FALSE}) Whether to print messages}

\item{parallel}{(Default: \code{FALSE}) If FALSE, no parallelization.
If TRUE, parallel execution using
\code{\link[BiocParallel:bpparam]{BiocParallel}}, see next argument
\code{BPPARAM}.}

\item{BPPARAM}{(Default: 
\code{\link[BiocParallel:bpparam]{bpparam()}})
An optional
parameter object passed internally to
\code{\link[BiocParallel:bplapply]{bplapply}}
when \code{parallel = TRUE}.
If not specified, the parameters last registered with register will be used.}

\item{...}{Additional parameters passed to
\code{\link[DESeq2:DESeq]{DESeq()}}}
}
\value{
A data.frame
}
\description{
This internal function runs a differential gene expression analysis using
DESeq2 (See \code{\link[DESeq2:DESeq]{DESeq}} for more detailed information).
Here, the surrogate variables found by \code{\link{run_sva}},
if any, will be added to the \code{\link[DESeq2:DESeqDataSet]{DESeqDataSet}}
before running the analysis.
}
\keyword{internal}
