% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_gsea.R
\name{run_dexseq}
\alias{run_dexseq}
\title{Run DEXSeq analysis}
\usage{
run_dexseq(
  dds,
  group_col,
  comparison,
  experiment_title = "MyLittleExperiment",
  store_results = FALSE,
  quiet = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{dds}{A DESeqDataSet}

\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{comparison}{The comparison to use for this particular experiment. Format example: "1v2"}

\item{experiment_title}{Title of your experiment. Your results will be stored in paste0("results/", experiment_title, "_pairedGSEA.RDS").}

\item{store_results}{(Default: TRUE) A logical indicating if results should be stored in the folder "results/".}

\item{quiet}{(Default: FALSE) Whether to print messages}

\item{parallel}{(Default: FALSE) If FALSE, no parallelization. If TRUE, parallel execution using BiocParallel, see next argument BPPARAM.}

\item{BPPARAM}{(Default: \code{BiocParallel::bpparam()}) An optional parameter object passed internally to bplapply when parallel = TRUE. If not specified, the parameters last registered with register will be used.}
}
\description{
Run DEXSeq analysis
}
\keyword{internal}
