% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_diff.R
\name{run_dexseq}
\alias{run_dexseq}
\title{Run DEXSeq analysis}
\usage{
run_dexseq(
    dds,
    group_col,
    baseline,
    case,
    experiment_title = "Experiment Title",
    store_results = FALSE,
    quiet = FALSE,
    parallel = FALSE,
    BPPARAM = BiocParallel::bpparam()
    )
}
\arguments{
\item{dds}{A DESeqDataSet. See \code{\link[DESeq2]{DESeqDataSet}} 
for more information about the object type.}

\item{group_col}{The metadata column specifying the what group each
sample is associated with}

\item{baseline}{Group value of baseline samples}

\item{case}{Group value of case samples}

\item{experiment_title}{Title of your experiment. Your results will be
stored in \code{paste0("results/", experiment_title, "_pairedGSEA.RDS")}.}

\item{store_results}{(Default: \code{FALSE})
A logical indicating if results should
be stored in the folder \code{"results/"}.}

\item{quiet}{(Default: \code{FALSE}) Whether to print messages}

\item{parallel}{(Default: \code{FALSE}) If FALSE, no parallelization.
If TRUE, parallel execution using
\code{\link[BiocParallel:bpparam]{BiocParallel}}, see next argument
\code{BPPARAM}.}

\item{BPPARAM}{(Default: 
\code{\link[BiocParallel:bpparam]{bpparam()}})
An optional
parameter object passed internally to
\code{\link[BiocParallel:bplapply]{bplapply}}
when \code{parallel = TRUE}.
If not specified, the parameters last registered with register will be used.}
}
\value{
A data.frame
}
\description{
This internal function runs a differential transcript usage analysis using
DEXSeq (See \code{\link[DEXSeq:DEXSeq]{DEXSeq}}
for more detailed information).
Here, the surrogate variables found by \code{\link{run_sva}}, if any,
will be added to the \code{\link[DEXSeq:DEXSeqDataSet]{DEXSeqDataSet}}
before running the analysis.
}
\keyword{internal}
