% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_diff.R
\name{run_sva}
\alias{run_sva}
\title{Run SVA on DESeqDataSet}
\usage{
run_sva(dds, quiet = FALSE)
}
\arguments{
\item{dds}{A DESeqDataSet. See \code{?DESeq2::DESeqDataSet} 
for more information about the object type.}

\item{quiet}{(Default: FALSE) Whether to print messages}
}
\value{
A DESeqDataSet
}
\description{
This internal function runs a surrogate variable analysis on the
count matrix in the DESeqDataSet. 
The found surrogate variables will then be added to the metadata
and the design formula in the DESeqDataSet object to be used in
the DGE and DTU analyses.
}
\keyword{internal}
