% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_gsea.R
\name{run_sva}
\alias{run_sva}
\title{Run SVA on DESeqDataSet}
\usage{
run_sva(dds, group_col, quiet = FALSE)
}
\arguments{
\item{dds}{A DESeqDataSet}

\item{group_col}{The metadata column specifying the what group each sample is associated with}

\item{quiet}{(Default: FALSE) Whether to print messages}
}
\description{
Run SVA on DESeqDataSet
}
\keyword{internal}
