% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Run DESeq2}
\usage{
runDESeq2(
  txCount,
  metadata,
  groupCol,
  comparison,
  design,
  preFilter = 10,
  parallel = FALSE,
  cores = 4
)
}
\description{
Run DESeq2
}
