% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Run DESeq2}
\usage{
runDESeq2(
  dds,
  groupCol,
  comparison,
  fitType = "local",
  tpm = FALSE,
  samples = NULL,
  dds_out = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\description{
Run DESeq2
}
