% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Run DESeq2}
\usage{
runDESeq2(
  md,
  archs4db,
  groupCol,
  comparison,
  tpm = FALSE,
  samples = "id",
  prefilter = 10,
  dds_out = FALSE,
  parallel = FALSE,
  cores = 4
)
}
\description{
Run DESeq2
}
