% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{runExperiment}
\alias{runExperiment}
\title{Run DESeq2 and DEXSeq analyses}
\usage{
runExperiment(
  row,
  archs4db,
  groupCol = "group_nr",
  tpm = TRUE,
  prefilter = 10,
  parallel = TRUE
)
}
\description{
Run DESeq2 and DEXSeq analyses
}
